model voc fixing

+++ Code/be/be/services/UserService.py
@@ -11,31 +11,38 @@ def username_exists(username):
 
 # Lấy tất cả người dùng
 def get_all_users():
-    return User.objects.all()
+    users = User.objects.all()
+    return {"errCode": 0, "message": "Success", "data": list(users.values())}
 
 
 # Lấy người dùng theo ID
 def get_user_by_id(user_id):
-    return get_object_or_404(User, id=user_id)
+    user = User.objects.filter(id=user_id).first()
+    if not user:
+        return {"errCode": 1, "message": "User not found"}
+    return {"errCode": 0, "message": "Success", "data": user}
 
 
 # Tạo mới người dùng
 def create_user(data):
     if "username" in data and username_exists(data["username"]):
-        raise ValidationError("username already exists.")
+        return {"errCode": 1, "message": "Username already exists"}
 
     if "password" in data:
         data["password"] = make_password(data["password"])  # Hash mật khẩu
 
-    return User.objects.create(**data)
+    user = User.objects.create(**data)
+    return {"errCode": 0, "message": "User created successfully", "data": user}
 
 
 # Cập nhật người dùng
 def update_user(user_id, data):
-    user = get_user_by_id(user_id)
+    user = User.objects.filter(id=user_id).first()
+    if not user:
+        return {"errCode": 1, "message": "User not found"}
 
     if "username" in data and username_exists(data["username"]) and user.username != data["username"]:
-        raise ValidationError("username already exists.")
+        return {"errCode": 2, "message": "Username already exists"}
 
     if "password" in data:
         data["password"] = make_password(data["password"])  # Hash mật khẩu mới
@@ -43,13 +50,17 @@ def update_user(user_id, data):
     for key, value in data.items():
         setattr(user, key, value)  # Gán giá trị mới
     user.save()
-    return user
+    return {"errCode": 0, "message": "User updated successfully", "data": user}
 
 
 # Xóa người dùng
 def delete_user(user_id):
-    user = get_user_by_id(user_id)
+    user = User.objects.filter(id=user_id).first()
+    if not user:
+        return {"errCode": 1, "message": "User not found"}
+
     user.delete()
+    return {"errCode": 0, "message": "User deleted successfully"}
 
 
 from django.contrib.auth import authenticate
